Updated room class to spawn enemies
Room class now is capable to spawn enemies in given amount in random places in the room. The enemies are child objects of the room. As for now the spawned objects' script is disabled, so they are not moving on the start of the game. I'm now implementing activating the enemies once i enter the room.