Door close + Unity version update + transitions
Now when player enters the room the door close until all enemies are killed, playing an animation for closing the door.

When all enemies are killed the door open, again with an animation.

Also updated Unity version to 2019.4.0f1 and it fixed some of the weird graphical glitches that were present before.

(bonus) Added transitions when the game first loads and when you die, then when it loads again.